+++ lib/recipes/anthropology/recipe.rb
@@ -60,5 +60,6 @@ ANTHROPOLOGY_RECIPE = Kitchen::BookRecipe.new(book_short_name: :anthropology) do
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end