+++ lib/recipes/biology/recipe.rb
@@ -71,5 +71,6 @@ BIOLOGY_RECIPE = Kitchen::BookRecipe.new(book_short_name: :biology) do |doc, _re
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end