+++ lib/recipes/english-composition/recipe.rb
@@ -75,6 +75,7 @@ do |doc, _resources|
   BakeLinkPlaceholders.v1(book: book)
   BakeToc.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 
   # Bake Custom Sections