+++ lib/recipes/finance/recipe.rb
@@ -124,5 +124,6 @@ FINANCE_RECIPE = Kitchen::BookRecipe.new(book_short_name: :finance) do |doc, _re
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end