+++ lib/recipes/marketing/recipe.rb
@@ -168,4 +168,5 @@ MARKETING_RECIPE = Kitchen::BookRecipe.new(book_short_name: :marketing) do |doc,
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
 end