+++ lib/recipes/neuroscience/recipe.rb
@@ -37,5 +37,10 @@ NEUROSCIENCE_RECIPE = Kitchen::BookRecipe.new(book_short_name: :neuroscience) do
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
+
+  note_classes = %w[meet-author]
+  BakeAutotitledNotes.v1(book: book, classes: note_classes)
+
 end