+++ lib/recipes/u-physics/recipe.rb
@@ -106,5 +106,6 @@ U_PHYSICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :uphysics) do |doc, 
   BakeLinkPlaceholders.v1(book: book)
   BakeUnitTitle.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end